+++ tests/__main__.py
@@ -30,6 +30,8 @@ if sys.version_info.minor > 5:
     from .test_dumpload import *
 if sys.version_info.minor >= 7:
     from .test_dataclass import *
+if sys.version_info.minor >= 8:
+    from .test_literal import *
 from .test_legacytuples_dataloader import *
 from .test_typechecks import *
